do refactoring

+++ BlazorViewer/Server/Services/DesignFileStorageService.cs
@@ -1,65 +1,78 @@
 ﻿
+using BlazorViewer.Server.Dtos;
+using BlazorViewer.Server.Options;
+using Microsoft.Extensions.Options;
+
 namespace BlazorViewer.Server.Services
 {
-    public class DesignFileStorageService : IProtoStorageService
+    public class DesignFileStorageService : IDesignStorageService
     {
-        private readonly IConfiguration _configuration;
+        private readonly FileStorageOptions _options;
         private readonly INameGeneratorService _nameGeneratorService;
 
-        public string StoragePath
-        {
-            get
-            {
-                return _configuration["FileStoragePath"];
-            }
-        }
-
-        public string FileExtension
-        {
-            get
-            {
-                return _configuration["StorageFileExtension"];
-            }
-        }
-
         public DesignFileStorageService(
-            IConfiguration configuration,
+            IOptions<FileStorageOptions> options,
             INameGeneratorService nameGeneratorService)
         {
-            _configuration = configuration;
+            _options = options.Value;
             _nameGeneratorService = nameGeneratorService;
         }
 
-        public void Create(Stream data)
+        public DesignDto UploadDesign(Stream content)
         {
             string filename = _nameGeneratorService.Generate("design");
-            CreateFile(filename, data);
+
+            CreateFile(filename, content);
+
+            return new DesignDto()
+            {
+                Name = filename
+            };
         }
 
-        public Stream Retrieve(string name)
+        public DesignDto GetDesignInfo(string name)
+        {
+            string path = GetPath(name);
+            var result = GetFileInfo(path);
+
+            return result;
+        }
+
+        public Stream GetDesignContent(string name)
         {
             string path = GetPath(name);
 
-            return RetrieveFromPath(path);
+            return GetFromPath(path);
         }
 
-        public IEnumerable<string> Retrieve()
+        public IEnumerable<DesignDto> ListDesigns()
         {
-            return GetFileNames(StoragePath, FileExtension);
+            return GetFileInfos();
         }
 
-        public void Update(string name, Stream data)
+        public DesignDto UpdateDesign(string name, Stream content)
         {
-            CreateFile(name, data);
+            CreateFile(name, content);
+
+            return new DesignDto()
+            {
+                Name = name
+            };
         }
 
-        public void Delete(string name)
+        public void DeleteDesign(string name)
         {
             string path = GetPath(name);
+
+            if(File.Exists(path) == false)
+            {
+                throw new FileNotFoundException();
+            }
+
             File.Delete(path);
         }
 
-        private Stream RetrieveFromPath(string path)
+        private Stream GetFromPath(string path)
         {
             MemoryStream result = new MemoryStream();
             using (FileStream input = File.OpenRead(path))
@@ -71,33 +84,42 @@ namespace BlazorViewer.Server.Services
             return result;
         }
 
-        private void CreateFile(string name, Stream data)
+        private void CreateFile(string name, Stream content)
         {
             string path = GetPath(name);
 
-            Directory.CreateDirectory(StoragePath);
+            Directory.CreateDirectory(_options.Path);
             using (FileStream output = File.Create(path))
             {
-                data.CopyTo(output);
+                content.CopyTo(output);
             }
         }
 
         private string GetPath(string filename)
-        {
-            return $"{StoragePath}{filename}{FileExtension}";
+        {           
+            return Path.Combine(_options.Path, $"{filename}{_options.FileExtension}");
         }
 
-        private IEnumerable<string> GetFileNames(
-            string directoryPath, 
-            string fileExtension)
+        private IEnumerable<DesignDto> GetFileInfos()
         {
             return Directory
-                .EnumerateFiles(directoryPath, $"*{fileExtension}")
-                .Select(
-                    x => x
-                    .Replace(FileExtension, "")
-                    .Split('/')
-                    .Last());
+                .EnumerateFiles(_options.Path, $"*{_options.FileExtension}")
+                .Select(x => GetFileInfo(x));
+        }
+
+        private DesignDto GetFileInfo(string path)
+        {
+            if (File.Exists(path) == false)
+            {
+                throw new FileNotFoundException();
+            }
+
+            string name = Path.GetFileNameWithoutExtension(path);
+
+            return new DesignDto()
+            {
+                Name = name
+            };
         }
     }
 }
